requirements.txt
faker
ipython
locustio
open-discussions-client
praw==4.5.1
